+++ src/objCmd.cpp
@@ -15,14 +15,14 @@ bool OBJCmd::registerCmd(void *obj, char *name, setFunc set, getFunc get)
   return ok;
 }
 
-void OBJCmd::set(char* name, byte* toSet, byte n)
+void OBJCmd::set(char* name, int* toSet, byte n)
 {
   for (byte i = 0; i < mNCMD; i++) //look for the cmd
     if (strcmp(name, mCmd[i].name)==0)
       (*mCmd[i].set)(mCmd[i].obj, toSet, n);
 }
 
-byte OBJCmd::get(char* name, byte* toGet)
+byte OBJCmd::get(char* name, int* toGet)
 {
   for (byte i = 0; i < mNCMD; i++) //look for the cmd
     if (strcmp(name, mCmd[i].name)==0)
